+++ src/unix.rs
@@ -1,10 +1,9 @@
-// WhoAmI
-// Copyright © 2017-2021 Jeron Aldaron Lau.
+// Copyright © 2017-2022 The WhoAmI Contributors.
 //
 // Licensed under any of:
-//  - Apache License, Version 2.0 (https://www.apache.org/licenses/LICENSE-2.0)
-//  - MIT License (https://mit-license.org/)
-//  - Boost Software License, Version 1.0 (https://www.boost.org/LICENSE_1_0.txt)
+// - Apache License, Version 2.0 (https://www.apache.org/licenses/LICENSE-2.0)
+// - Boost Software License, Version 1.0 (https://www.boost.org/LICENSE_1_0.txt)
+// - MIT License (https://mit-license.org/)
 // At your choosing (See accompanying files LICENSE_APACHE_2_0.txt,
 // LICENSE_MIT.txt and LICENSE_BOOST_1_0.txt).
 
@@ -348,7 +347,7 @@ pub fn hostname() -> String {
     string_from_os(hostname_os())
 }
 
-pub fn hostname_os() -> OsString {
+fn hostname_os() -> OsString {
     // Maximum hostname length = 255, plus a NULL byte.
     let mut string = Vec::<u8>::with_capacity(256);
     unsafe {
@@ -562,7 +561,7 @@ pub fn lang() -> impl Iterator<Item = String> {
         .next()
         .unwrap_or("en_US")
         .to_string()
-        .replace("_", "-");
+        .replace('_', "-");
     LangIter {
         array,
         index: Some(true),